+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - `:all` can now be used to have an Object persist for all frames of an animation
+- added support for Jupyter notebooks
 
 ## v0.4.0 (9th of January 2021)
 - added ImageIO and ImageMagick as dependencies